access check and time zone
*  Setting timezone in a function.
* Changed start date and end date types from String to Number
* Checking access to the database on all requests.
* Showing a solution to a user, in case he doesn't have access to a database.

+++ src/components/BodyTableRemoveModel/index.tsx
@@ -5,6 +5,10 @@ import { IoMdClose } from "react-icons/io";
 
 import { useInfoContext } from "../../contexts/userInfoContext";
 import { deleteModelDatabase } from "../../database/firebase";
+import {
+  accessDenied,
+  checkAuthorizations,
+} from "../../helpers/authorizations";
 
 type Props = {
   item: ProductDatabase;
@@ -22,6 +26,13 @@ function BodyTableRemoveModel(props: Props) {
     const user = state.infoUser?.email;
     const token = await state.infoUser?.getIdToken();
     const authorizedDatabase = state.databaseAuth;
+    const infoUser = state;
+    const authorized = await checkAuthorizations(infoUser);
+
+    if (!authorized && authorizedDatabase !== null) {
+      alert(accessDenied(infoUser));
+      return;
+    }
 
     await deleteModelDatabase(item.id, user, token, authorizedDatabase);
     getProducts();
