+++ src/components/TableItem/index.tsx
@@ -5,6 +5,7 @@ import * as C from "./styles";
 import { useInfoContext } from "../../contexts/userInfoContext";
 import { IoMdClose } from "react-icons/io";
 import { deleteTransactionDatabase } from "../../database/firebase";
+import { checkAuthorizations } from "../../helpers/authorizations";
 
 type Props = {
   item: ItemDataBase;
@@ -24,6 +25,15 @@ function TableItem(props: Props) {
     const user = state.infoUser?.email;
     const token = await state.infoUser?.getIdToken();
     const authorizedDatabase = state.databaseAuth;
+    const infoUser = state;
+    const authorized = await checkAuthorizations(infoUser);
+
+    if (!authorized && authorizedDatabase !== null) {
+      alert(
+        `Não autorizado a acessar ${infoUser.databaseAuth} \nVá para configurações e altere o banco de dados.`
+      );
+      return;
+    }
 
     await deleteTransactionDatabase(itemId, user, token, authorizedDatabase);
